### 2021-10-20 #147 @whedon
[PRE REVIEW]: PCP Notebooks: A Preparation Course for Python with a Focus on Signal Processing
**Submitting author:** @sebastianrosenzweig (<a href="http://orcid.org/0000-0003-4964-9217">Sebastian Rosenzweig</a>)
**Repository:** <a href="https://github.com/meinardmueller/PCP" target ="_blank">https://github.com/meinardmueller/PCP</a>
**Version:** v1.0.3
**Managing EiC:** Jordan Gorzalski

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
