### 2021-11-01 #147 @moorepants
@boisgera and @alfredocarella, thank you both for volunteering to review this submission. The timeline works for us, so we can plan to have the initial reviews by November 26th. Let me know if that doesn't work. Please have a look at the JOSE reviewing guide: https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html to get an introduction for reviewing here or to refresh your memory if you have already reviewed with us.

I will add you all as reviewers and new issue will be opened where the review will occur.

Let me know if you have any questions.
