### 2021-11-01 #148 @whedon
[REVIEW]: PCP Notebooks: A Preparation Course for Python with a Focus on Signal Processing
**Submitting author:** @sebastianrosenzweig (<a href="http://orcid.org/0000-0003-4964-9217">Sebastian Rosenzweig</a>)
**Repository:** <a href="https://github.com/meinardmueller/PCP" target ="_blank">https://github.com/meinardmueller/PCP</a>
**Version:** v1.1.2
**Reviewer:** @boisgera , @alfredocarella

- [x] #153
