### 2021-11-11 #148 @alfredocarella
@moorepants that is correct. The mentioned invite link (quote below) expired after 7 days.
I can no longer access https://github.com/openjournals/jose-reviews/invitations to accept the role as reviewer

> ## Reviewer instructions & questions
>
> @boisgera & @alfredocarella, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
>
>     1. Make sure you're logged in to your GitHub account
>
>     2. Be sure to accept the invite at this URL: https://github.com/openjournals/jose-reviews/invitations
