### 2021-11-23 meinardmueller/PCP#6 @boisgera
String Formatting
In the recent versions of the Python documentation, the options that are emphasized for string formatting are:

 1. formatted string literal (f-strings),
 2. `str.format` method,
 3. string slicing and concatenation

The printf-style %-based method is only briefly mentionned and referred to as "old-style formatting". (See https://docs.python.org/3/tutorial/inputoutput.html#formatted-string-literals.)

While the feature is not formally deprecated, I'd rather have students learn f-strings instead of printf-style formatting, as in:

    >>> string = f'x = {x}'

(I would not use the recent syntax `string = f'{x = }'` in such a course, but YMMV). Of course, since this preparation course is not meant to be comprehensive, if the f-string approach is considered too heavy, the 3rd option (concatenation) could be an option too:

    >>> string = 'x = ' + str(x)
