### 2021-11-23 meinardmueller/PCP#7 @boisgera
Matrices, Vectors and Arrays
Unit 3 states that row vectors and columns vectors are special cases of matrices (with either 1 column or 1 row). While this is also how I see things, however idiomatic NumPy code would probably represent matrices as 2D arrays but row and column vectors as 1D arrays. The same object can then be interpreted and used as a row or column vector in different contexts.

Instead of

    >>> A = np.array([[0, 1], [2, 3]])
    >>> x = np.array([[4], [5]])
    >>> y = np.array([[6, 7]])
    >>> np.dot(A, x)
    array([[ 5],
           [23]])
    >>> y.dot(A)
    array([[14, 27]])
    >>> np.dot(y, A)
    array([[14, 27]])
    >>> np.dot(y, x)
    array([[59]])
    >>> np.dot(x, y)
    array([[24, 28],
           [30, 35]])

one would generally do

    >>> A = np.array([[0, 1], [2, 3]])
    >>> x = np.array([4, 5])
    >>> y = np.array([6, 7])
    >>> np.dot(A, x)
    array([ 5, 23])
    >>> np.dot(y, A)
    array([14, 27])
    >>> np.dot(y, x)
    59
    >>> np.outer(x, y)
    array([[24, 28],
           [30, 35]])

I'd rather see that latter kind of idiom used (and have the solution of exercise 2 updated accordingly).
