### 2021-11-23 #148 @boisgera
## Issues

I have noticed a small number of typos and made a PR [here](https://github.com/meinardmueller/PCP/pull/5).

I have also described some small issues here :

  - https://github.com/meinardmueller/PCP/issues/6
  - https://github.com/meinardmueller/PCP/issues/7

but this is mostly some nitpicking, and should not been considered show-stoppers (I would be ok with a "won't fix").

## Review

The content provided by the authors is nicely structured, the scope of the content consistent with the stated aim
and the mix of Python introduction and audio / signal processing works really well in my humble opinion. More generally many aspects of the material (notebooks formatting, scope, length and balance, learning objectives, exercices, documentation wrt execution, etc.) demonstrate that a great care has been taken to design the material and make it useful.

I have initially been surprised by the large amount of raw HTML in markdown cells (for example hyperlinks are frequently given in HTML form) and then I noticed the use of bootstrap classes to provide a better styling of the content cells, where HTML is indeed required. While I'd rather have more markdown and less HTML (less technical for authors), the styling is undeniably useful (for example to easily spot exercises with a different color), so the option selected actually makes sense.

With respect to the checklist, I find only one thing that is lacking at this point (unless I am mistaken) :

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support

I feel that the GitHub issue system is sufficiently known to consider that 2) is covered. IMHO, if the discussions features of github was activated and a brief mention of it was made in the README, that would also cover 3). The status of 1) could be explicited in the Acknowledgements section of the README (for example).

Once this point has been resolved, I recommend the submission to be accepted without any reservation. And I'd like to thank the authors for this contribution!
