### 2021-11-24 #148 @sebastianrosenzweig
@boisgera many thanks for your thorough review and the pull requests! We are happy that our work is well appreciated.
@moorepants We will implement the changes in the next few days. Also we'll make clearer in the README.md how to contribute, report issues, or seek support.
