### 2021-12-01 #148 @alfredocarella
@moorepants I am just beginning to go through the content after the preliminary checks. It looks very well so far. I will submit my comments as soon as I have them
