### 2021-12-04 #148 @alfredocarella
## Comments:
- I have submitted 2 pull requests [here](https://github.com/meinardmueller/PCP/pull/9) and [here](https://github.com/meinardmueller/PCP/pull/8), suggesting fixes for small typos and some broken links. These are minimal issues, but I believe the changes reflect what the authors intended to do.

- As mentioned by the other reviewer, a very brief paragraph with guidelines for contributing (or eventually a statement that contributions are not expected) would be nice.

## Suggestions (optional, and subjective)
- In my humble and subjective opinion, using f-strings consistently throughout the course might be the most intuitive approach for students.

- It might be helpful for users to add some numbering system to the notebook file names (for example "PCP_01_getstarted.ipynb"). This way a Jupyter notebook user will not need to revisit "PCP.ipynb" to find out which notebook in the series is next.

## Overall review
The selection and treatment of topics is very good. The notebook content achieves the outcome of providing tangible and concrete examples that feel (and are) useful, without overwhelming students with details. Moreover the consistent style and formatting reduce the cognitive load by giving visual cues that help in navigating the notebooks quickly and effectively. The exercises are challenging enough to require a bit of time, but not so much as to be discouraging.
I confidently recommend this submission for publication, as I am sure it will be helpful for many.
