### 2021-12-06 995e421
small format fix, added contributing statement
