+++ README.md
@@ -38,6 +38,9 @@ One can also use [Binder](https://mybinder.org/) to execute the PCP notebooks. T
 
 https://mybinder.org/v2/gh/meinardmueller/PCP/master
 
+## Contributing
+We are happy for suggestions and contributions. However, to facilitate the synchronization, we would be grateful for either directly contacting us via email (meinard.mueller@audiolabs-erlangen.de) or for creating an issue](https://github.com/meinardmueller/PCP/issues) in our GitHub repository. Please do not submit a pull request without prior consultation with us.
+
 ## Acknowledgements
 
 We want to thank the various people who have contributed to the design, implementation, and code examples of the notebooks. We mention the main contributors in alphabetical order: Michael Krause, Heinrich Löllmann, Meinard Müller, Sebastian Rosenzweig, Frank Zalkow. The [International Audio Laboratories Erlangen](https://www.audiolabs-erlangen.de/) are a joint institution of the [Friedrich-Alexander-Universität Erlangen-Nürnberg (FAU)](https://www.fau.eu/) and [Fraunhofer Institute for Integrated Circuits IIS](https://www.iis.fraunhofer.de/en.html).
