### 2021-12-06 #148 @sebastianrosenzweig
Dear @moorepants, @boisgera, @alfredocarella,

Thank you very much for your helpful feedback!

We took some time to incorporate your comments and improve the material. Here is a summary of the changes we made to the `paper` branch:
* Added a "Contributing" statement to the README.md
* Fixed typos and grammar
* Replaced the old-fashioned string formatting with f-Strings consistently
* Numbered notebooks for better orientation
* Added a section "HTML Export" to the notebook PCP_01_getstarted.ipynb, where commands for exporting the PCP notebooks are provided
* Added HTML exports to the repository to avoid invalid HTML links

We hope that the material is now ready for publishing. Thanks again for taking the time to review our work!
