### 2021-12-08 #148 @alfredocarella
@moorepants The authors have satisfactorily addressed all of my comments on the `paper` branch. I have thus checked the last item on my list.
@sebastianrosenzweig It might be a good idea to include the changes on `master` as well, particularly the "Contributing" statement in `README.md`.
