### 2021-12-08 #148 @sebastianrosenzweig
@alfredocarella Yes, we will merge the `paper` branch into the `master` branch as soon as the review process is completed.
