### 2021-12-09 #148 @boisgera
> @boisgera and @alfredocarella Can you all review the updates and confirm if it meets your expectations?
> > @moorepants The authors have satisfactorily addressed all of my comments on the paper branch. I have thus checked the last item on my list.
> >@sebastianrosenzweig It might be a good idea to include the changes on master as well, particularly the "Contributing" statement in README.md.

@moorepants Yes, the updates have nicely adressed my feedback. I have checked the last item of the list.
