### 2021-12-15 #148 @moorepants
@sebastianrosenzweig I've checked over the reviewer's checklists, comments, and your adjustments as well as going through the submitted paper and modules. I agree with the two reviewers and I recommend this paper and repository for publishing. I believe the next step will be for you to make a release of the new version and then publish that release to Zenodo as an archive. Once you have the Zenodo DOI you can provide that here so I can update it. Congrats on the nice work! It is a great reusable addition to computational education resources. Thanks to @boisgera and @alfredocarella for their careful and thoughtful reviews.
