+++ data/PCP_bibtex_MuellerR21_PCP_JOSE.txt
@@ -4,7 +4,7 @@
   journal   = {Journal of Open Source Education ({JOSE})},
   volume    = {},
   number    = {},
-  year      = {},
+  year      = {2021},
   pages     = {},
   doi       = {},
   url-demo = {https://github.com/meinardmueller/PCP}
