### 2021-12-16 #148 @sebastianrosenzweig
Dear @moorepants,
We are delighted that our submission is accepted for publication.
We just merged the `paper` branch into the `master`, made a new release and published it on Zenodo.
The DOI is: 10.5281/zenodo.5786628
Thanks again for your time.
