### 2021-12-16 #148 @moorepants
@openjournals/jose-eics I'm not sure why the generate pdf passes but the recommend accept fails to build the pdf.
