### 2021-12-16 #148 @labarba
@openjournals/dev – Can you have a look at this JOSE submission? The command `@whedon recommend-accept` fails to compile the PDF (but `generate pdf` does work).
