### 2021-12-23 f2f06e3
revert paper changes
