### 2021-12-23 #148 @sebastianrosenzweig
Sorry for the spam, I just checked whether this could be a formatting issue of the pdf, but no insights here... I reverted my changes such that the pdf builds again.
