### 2021-12-23 #148 @moorepants
@sebastianrosenzweig I found this: https://github.com/openjournals/joss-reviews/issues/3326#issuecomment-883273118 you can see that the PR to the library edited the metadata file to change something about how the bib file was selected. Can you check that out and see if the fix relates to your metadata?
