### 2021-12-23 8709f15
fixed bibliography markdown formatting
