+++ paper/paper.md
@@ -15,13 +15,8 @@ affiliations:
  - name: International Audio Laboratories Erlangen
    index: 1
 date: 01 November 2021
-bibliography:
-  - paper.bib
+bibliography: paper.bib
 
-# Optional fields if submitting to a AAS journal too, see this blog post:
-# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
-# aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
-# aas-journal: Astrophysical Journal <- The name of the AAS journal.
 ---
 
 # Summary
