### 2021-12-23 #148 @moorepants
I think that you'll need to make a new version and zenodo upload so that it captures the correct version.
