### 2021-12-23 #148 @sebastianrosenzweig
Would it be possible to proceed without creating a new Zenodo version?
I would like to avoid rolling out a new version that has no noticible change in the actual code or paper. This also did not seem to be necessary in the previously mentioned publication [3326#issuecomment-883273118l](https://github.com/openjournals/joss-reviews/issues/3326#issuecomment-883273118l). In our case, the notebooks need to be updated anyways after publication of the article, since they reference to the JOSE article pdf.

Since I don't have direct access to Zenodo and due to the approaching holiday season, we would have to wait until next year to roll out a new Zenodo version.
