### 2021-12-23 #148 @moorepants
My understanding is that we want the Zenodo archive to match the paper version we publish.  In this instance, this is a very minor difference in the Zenodo archive and the repository commit that produces the paper (slight difference in the meta data file), but it is very simple to tag a new release and push a new archive. That ensures the published paper exactly matches the archive.

> In our case, the notebooks need to be updated anyways after publication of the article, since they reference to the JOSE article pdf.

If your notebooks reference the publication before it is published, that creates a circular reference. You can update the repo all you want after publication, but the published version is the zenodo archive not the every changing repo.
