### 2021-12-23 #148 @sebastianrosenzweig
Ok, so the new version is v1.1.1 which has the following DOI on Zenodo: 10.5281/zenodo.5802031
