### 2021-12-23 #148 @labarba
hi @sebastianrosenzweig 👋  — Could you edit the metadata of the [Zenodo archive](https://zenodo.org/record/5802031), so the title and author list match the JOSE paper? Zenodo pulls these automatically from the repo and includes every committer by default.
