### 2021-12-23 #148 @labarba
I also notice that you upped the version to `v1.1.1`. Should we change the paper's linked version to that?
