### 2021-12-23 #148 @labarba
Quoting @moorepants:
> If your notebooks reference the publication before it is published, that creates a circular reference. You can update the repo all you want after publication, but the published version is the zenodo archive not the ever changing repo.

This can be done because JOSE will create a DOI based on the review-thread number, so we already know what the DOI of the published paper will be: `10.21105/jose.00138`
