### 2022-01-04 #148 @moorepants
> I also notice that you upped the version to v1.1.1. Should we change the paper's linked version to that?

Yes, I'm not sure how to do this though.
