### 2022-01-04 #148 @moorepants
@sebastianrosenzweig, you still need to address this comment:

> hi @sebastianrosenzweig wave — Could you edit the metadata of the Zenodo archive, so the title and author list match the JOSE paper? Zenodo pulls these automatically from the repo and includes every committer by default.
