### 2022-01-04 f838d7f
updated pdf, references, links, and version
