### 2022-01-04 #148 @sebastianrosenzweig
Dear @moorepants,
First of all Happy New Year and sorry, for the delay in the process.
We had to update the notebooks again to fix some links, update the year and references. Also we published a new version 1.1.2 .
Could you please use v1.1.2 for the article? The Zenodo DOI is: 10.5281/zenodo.5817272
We also changed the Zenodo authors to match the JOSE authors.
Thanks!
