### 2022-01-04 #148 @moorepants
@openjournals/jose-eics, this should now be ready for full acceptance. The PR `#76` looks correct to me.
