### 2022-01-24 #148 @moorepants
@openjournals/jose-eics Hi, just checking on this again. It may only need the final approval. Hopefully all concerns are addressed.
