### 2022-01-24 #148 @labarba
Huge thanks to our Editor: @moorepants, and Reviewers: @boisgera , @alfredocarella — your contributions make this adventure in indie publishing a reality and we are grateful 🙏
